backend/requirements.txt
Django==5.1.5
djangorestframework==3.15.2
django-cors-headers==4.6.0
django-filter==24.3
channels==4.2.0
channels_redis==4.2.1
celery==5.4.0
daphne==4.1.2
PyJWT==2.10.1
pillow==11.1.0
python-decouple==3.8